+++ include/pcomb/privates/adaptive.h
@@ -73,7 +73,7 @@ class AdaptiveParser
       return ResultType();
     }
 
-    int consumed_number = result.get_consumed_number();
+    size_t consumed_number = result.get_consumed_number();
     return ResultType(consumed_number, AdaptedValue<P, F>::invoke(
         func_, std::move(result).get_value()));
   }
