+++ include/pcomb/privates/constructive.h
@@ -72,7 +72,7 @@ class ConstructiveParser : public Parser<typename P::CharType, T> {
       return ResultType();
     }
 
-    int consumed_number = result.get_consumed_number();
+    size_t consumed_number = result.get_consumed_number();
     return ResultType(consumed_number,
         ConstructedValue<T, P>::construct(std::move(result).get_value()));
   }
