+++ include/pcomb/privates/strict_alternative.h
@@ -49,7 +49,7 @@ class StrictAlternativeParser : public StrictAlternativeBaseType<P1, PS...> {
     static ResultType parse(const StorageType& parsers, StreamType* stream) {
       auto result = std::get<I>(parsers).parse(stream);
       if (result.success()) {
-        int consumed = result.get_consumed_number();
+        size_t consumed = result.get_consumed_number();
         return ResultType(consumed, ValueType(std::in_place_index<I>,
                                               std::move(result).get_value()));
       }
@@ -62,7 +62,7 @@ class StrictAlternativeParser : public StrictAlternativeBaseType<P1, PS...> {
     static ResultType parse(const StorageType& parsers, StreamType* stream) {
       auto result = std::get<StorageSize-1>(parsers).parse(stream);
       if (result.success()) {
-        int consumed = result.get_consumed_number();
+        size_t consumed = result.get_consumed_number();
         return ResultType(consumed,
                           ValueType(std::in_place_index<StorageSize-1>,
                                     std::move(result).get_value()));
